chore: introduce python feature
