+++ dozer-sql/Cargo.toml
@@ -16,7 +16,9 @@ uuid = {version = "1.3.0", features = ["v1", "v4", "fast-rng"]}
 dozer-types = {path = "../dozer-types"}
 dozer-core = {path = "../dozer-core"}
 dozer-tracing = {path = "../dozer-tracing"}
-pyo3 = {version = "0.18.1", features = ["auto-initialize"]}
 
 [dev-dependencies]
 tempdir = "0.3.7"
+
+[features]
+python = []
