+++ dozer-sql/src/pipeline/expression.rs
@@ -8,7 +8,9 @@ pub mod geo;
 pub mod logical;
 pub mod mathematical;
 pub mod operator;
-pub mod python_udf;
 pub mod scalar;
+
+#[cfg(feature = "python")]
+pub mod python_udf;
 #[cfg(test)]
 mod tests;
