+++ dozer-sql/src/pipeline/expression/python_udf.rs
@@ -3,9 +3,9 @@ use crate::pipeline::errors::PipelineError::UnsupportedSqlError;
 use crate::pipeline::errors::UnsupportedSqlError::GenericError;
 use crate::pipeline::expression::execution::{Expression, ExpressionExecutor};
 use dozer_types::ordered_float::OrderedFloat;
+use dozer_types::pyo3::types::PyTuple;
+use dozer_types::pyo3::Python;
 use dozer_types::types::{Field, FieldType, Record, Schema};
-use pyo3::types::PyTuple;
-use pyo3::Python;
 use std::env;
 use std::path::PathBuf;
